refactor: alternative header
fixed/refactored:
- storybook args cleaned up & comply with latest version
- remove JS-based animations in favor of CSS
- simplified implementation
- a cleanup of styles.ts
- rename things, export from `index.ts`
